G3.1: replaced tests to src | made action unit test
