--- frontend/__tests__/components/Account.test.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import Account from '../../components/Account';
-import { text } from 'stream/consumers';
-
-describe("/components/auth/Account.test.tsx", () => {
-  ['Настройки'].forEach(text => {
-    test('text "' + text+ '" not found', () => {
-      render(<Account />);
-      const element = screen.getByText(new RegExp(text, 'i'));
-      expect(element).toBeInTheDocument();
-    });
-  });
-});
